Move DirectService into its own crate
