+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
   "tower-balance",
   "tower-buffer",
+  "tower-direct-service",
   "tower-discover",
   "tower-filter",
   "tower-in-flight-limit",
