+++ tower-buffer/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 [dependencies]
 futures = "0.1"
 tower-service = { version = "0.1", path = "../tower-service" }
+tower-direct-service = { version = "0.1", path = "../tower-direct-service" }
 
 [dev-dependencies]
 tower-mock = { version = "0.1", path = "../tower-mock" }
