+++ tower-direct-service/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+
+name = "tower-direct-service"
+# When releasing to crates.io:
+# - Update html_root_url.
+# - Update CHANGELOG.md.
+# - Update documentation URL
+# - Create "v0.1.x" git tag.
+version = "0.1.0"
+authors = ["Carl Lerche <me@carllerche.com>", "Jon Gjengset <jon@thesquareplanet.com>"]
+license = "MIT"
+readme = "README.md"
+repository = "https://github.com/tower-rs/tower"
+homepage = "https://github.com/tower-rs/tower"
+documentation = "https://docs.rs/tokio-direct-service/0.1.0"
+description = """
+Trait representing an asynchronous, request / response based, service that must be driven.
+"""
+categories = ["asynchronous", "network-programming"]
+
+[dependencies]
+futures = "0.1.23"
